+++ ptn/missionalertbot/botcommands/DatabaseInteraction.py
@@ -4,8 +4,10 @@ A Cog for commands that are primarily concerned with the bot's databases.
 """
 
 # import libraries
+import aiohttp
 import asyncio
 import copy
+import os
 import re
 
 # import discord.py
@@ -16,14 +18,16 @@ from discord.ext import commands
 # local classes
 from ptn.missionalertbot.classes.CommunityCarrierData import CommunityCarrierData
 from ptn.missionalertbot.classes.NomineesData import NomineesData
-from ptn.missionalertbot.classes.Views import db_delete_View, BroadcastView, CarrierEditView
+from ptn.missionalertbot.classes.Views import db_delete_View, BroadcastView, CarrierEditView, MissionDeleteView
 
 # local constants
 import ptn.missionalertbot.constants as constants
-from ptn.missionalertbot.constants import bot, cmentor_role, admin_role, cteam_bot_channel, cteam_bot_channel, bot_command_channel, cc_role
+from ptn.missionalertbot.constants import bot, cmentor_role, admin_role, cteam_bot_channel, cteam_bot_channel, bot_command_channel, cc_role, bot_spam_channel, \
+    certcarrier_role, trainee_role, mission_command_channel
 
 # local modules
-from ptn.missionalertbot.database.database import find_nominee_with_id, carrier_db, CarrierDbFields, CarrierData, find_carrier, backup_database, add_carrier_to_database, find_carriers_mult, find_commodity, find_community_carrier, CCDbFields
+from ptn.missionalertbot.database.database import find_nominee_with_id, carrier_db, CarrierDbFields, CarrierData, find_carrier, backup_database, \
+     add_carrier_to_database, find_carriers_mult, find_commodity, find_community_carrier, CCDbFields, find_mission
 from ptn.missionalertbot.modules.helpers import on_app_command_error, check_roles, check_command_channel, _regex_alphanumeric_with_hyphens
 from ptn.missionalertbot.modules.Embeds import _add_common_embed_fields, _configure_all_carrier_detail_embed
 
@@ -48,6 +52,7 @@ findid - database
 findshort - database
 /info - general/database
 /owner - general/database
+
 """
 
 
@@ -156,7 +161,6 @@ class DatabaseInteraction(commands.Cog):
 
 
     # command to retrieve text for all nominations for a given user from the database
-    # TODO: ephemeral with option to broadcast
     @app_commands.command(name='cp_nomination_details', description='Shows details of all nominations for a given user by ID.')
     @app_commands.describe(userid='The Discord ID of the target user. Use Developer Mode to retrieve user IDs.')
     @check_roles([cmentor_role(), admin_role()])
@@ -647,7 +651,7 @@ class DatabaseInteraction(commands.Cog):
 
     @app_commands.command(name="info",
                           description="Private command: Use in a Fleet Carrier's channel to show information about it.")
-    async def _info(self, interaction: discord.Interaction):
+    async def info(self, interaction: discord.Interaction):
 
         print(f'/info command carrier_data called by {interaction.user} in {interaction.channel}')
 
@@ -686,7 +690,9 @@ class DatabaseInteraction(commands.Cog):
             embed = discord.Embed(title=f"Welcome to {carrier_data.carrier_long_name} ({carrier_data.carrier_identifier})", color=constants.EMBED_COLOUR_OK)
             embed = _add_common_embed_fields(embed, carrier_data, interaction)
             carrier_owner_obj = bot.get_user(carrier_data.ownerid)
-            thumbnail_file = discord.File(f"images/{carrier_data.carrier_short_name}.png", filename="image.png")
+            image_name = carrier_data.carrier_short_name + '.png'
+            image_path = os.path.join(constants.IMAGE_PATH, image_name)
+            thumbnail_file = discord.File(image_path, filename="image.png")
             embed.set_thumbnail(url="attachment://image.png")
             embed.set_author(name=carrier_owner_obj.name, icon_url=carrier_owner_obj.display_avatar)
             interaction.user = carrier_owner_obj
@@ -859,3 +865,33 @@ class DatabaseInteraction(commands.Cog):
                 return
         else:
             await ctx.send(f"No Community Carrier registered to {owner.display_name}")
+
+
+    # manually delete a mission from the database
+    @app_commands.command(name='admin_delete_mission', description='Manually remove a mission from the database.')
+    @app_commands.describe(carrier='Carrier name to search for in the missions database.')
+    @check_roles([admin_role()])
+    @check_command_channel(bot_command_channel())
+    async def admin_delete_mission(self, interaction: discord.Interaction, carrier: str):
+        print(f"admin_delete_mission called by {interaction.user.display_name} ({interaction.user.id})")
+        mission_data = find_mission(carrier, "carrier")
+        if not mission_data:
+            embed = discord.Embed(
+                description=f"**ERROR**: no trade missions found for carriers matching \"**{carrier}\"**.",
+                color=constants.EMBED_COLOUR_ERROR
+            )
+            return await interaction.response.send_message(embed=embed, ephemeral=True)
+        else:
+            embed = discord.Embed(
+                description=f"Please confirm you want to delete the mission for **{mission_data.carrier_name}**. "\
+                    "This should **only** be done if `/mission complete` and `/cco complete` will not work. " \
+                    "Deleting a mission this way **will** require manual cleanup of any remaining mission elements.",
+                    color=constants.EMBED_COLOUR_QU
+            )
+
+            view=MissionDeleteView(mission_data, interaction.user)
+
+            await interaction.response.send_message(embed=embed, view=view)
+
+            view.message = await interaction.original_response()
+